Update example to current standard

+++ public/usage-examples/sprites/sprite_x-1-example.cpp
@@ -1,5 +1,4 @@
 #include "splashkit.h"
-#include <string>  
 
 int main()
 {
@@ -11,14 +10,12 @@ int main()
     // Setting the x coordinate in reference to the window
     sprite_set_x(player_sprite, 300);
 
-    // Retrieve the x position using sprite_x function
+    // Retrieve the x position
     double x_position = sprite_x(player_sprite);
-    
+
     clear_screen(COLOR_BLACK);
     draw_sprite(player_sprite);
-
-    draw_text("Sprite X: " + std::to_string(x_position), COLOR_WHITE, 10, 10);
-
+    draw_text("Sprite X: " + std::to_string((int)x_position), COLOR_WHITE, 10, 10);
     refresh_screen();
     delay(5000);
